Implemented responsive layout for pages and decomposition components
