+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import RoutedApp from './App'
 import './index.css'
+import RoutedApp from './router/RoutedApp'
 import * as serviceWorker from './serviceWorker'
 
 import { Provider } from 'react-redux'
