+++ src/styles/components/AppFooter.css
@@ -3,26 +3,37 @@
     text-align: center;
 }
 
-.app-footer__social {
+.app-footer .container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
 
+.app-footer__social {
+    padding: 10px;
+}
+
+.app-footer__copyright {
+    margin: 0 30px 0 0;
 }
 
 .social {
-
+    font-size: 0;
 }
 
 .social__link {
-
+    cursor: pointer;
 }
 
-.social__link--fb {
-
+.social__link:hover {
+    opacity: 0.7;
 }
 
-.social__link--github {
-
+.social__link svg {
+    width: 30px;
+    height: 30px;
 }
 
-.social__link--telegram {
-
+.social__link:not(:last-of-type) {
+    margin-right: 10px;
 }
