+++ src/styles/components/AppNavigation.css
@@ -1,0 +1,52 @@
+
+.app-nav {
+    display: flex;
+    align-items: baseline;
+    justify-content: space-around;
+    padding: 10px 30px;
+}
+
+.app-nav__link {
+    margin-right: 50px;
+    font-size: 14px;
+    font-weight: bold;
+    color: rgba(0, 0, 0, 0.5);
+    text-decoration: none;
+    text-transform: uppercase;
+}
+
+.app-nav__link:hover {
+    color: rgba(0, 0, 0, 0.7);
+}
+
+.app-nav__link--active {
+    color: #000;
+}
+
+@media (max-width: 650px) {
+    .app-nav {
+        position: absolute;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+        background-color: #fff;
+        padding: 20px 15px;
+        flex-direction: column;
+        justify-content: flex-start;
+        align-items: center;
+        z-index: 1;
+        transition: transform 500ms ease-in-out;
+        transform: translateX(-100%);
+    }
+
+    .app-nav__link {
+        margin-right: 0;
+        margin-bottom: 20px;
+        font-size: 1.25em;
+    }
+
+    .app-nav--show {
+        transform: translateX(0);
+    }
+}
