+++ src/styles/pages/Home.css
@@ -51,3 +51,35 @@
         left: 200px;
     }
 }
+
+@media (min-width: 651px) {
+    .app-home__image--react {
+        top: 155px;
+        left: 40%;
+    }
+}
+
+@media (max-width: 650px) {
+    .app-home__content {
+        flex-direction: column;
+    }
+
+    .app-home__wrapper {
+        order: 2;
+    }
+
+    .app-home__wrapper--images {
+        margin-top: 20px;
+        margin-bottom: 20px;
+        order: 1;
+    }
+
+    .app-home__image {
+        position: static;
+        width: 40%;
+    }
+
+    .app-home__image--react {
+        margin-right: 30px;
+    }
+}
